refactor websocket dir structure, delete deprecated classes
