+++ src/main/java/com/trymad/litechess_monolith/chessgame/internal/service/ChessUtilService.java
@@ -3,7 +3,7 @@ package com.trymad.litechess_monolith.chessgame.internal.service;
 import java.util.List;
 
 import com.trymad.litechess_monolith.chessgame.api.model.GameMove;
-import com.trymad.litechess_monolith.websocket.MoveRequest;
+import com.trymad.litechess_monolith.websocket.api.dto.MoveRequest;
 
 public interface ChessUtilService {
