--- src/main/java/com/trymad/litechess_monolith/websocket/GameCreatedDTO.java
@@ -1,5 +1,0 @@
-package com.trymad.litechess_monolith.websocket;
-
-public record GameCreatedDTO(Long gameId) {
-	
-}
