+++ src/main/java/com/trymad/litechess_monolith/websocket/api/dto/CreateGameRequest.java
@@ -1,4 +1,4 @@
-package com.trymad.litechess_monolith.websocket;
+package com.trymad.litechess_monolith.websocket.api.dto;
 
 import com.trymad.litechess_monolith.chessgame.api.model.GameCategory;
 import com.trymad.litechess_monolith.chessgame.api.model.GameVariant;
