+++ src/main/java/com/trymad/litechess_monolith/websocket/api/dto/GameCreatedDTO.java
@@ -1,0 +1,5 @@
+package com.trymad.litechess_monolith.websocket.api.dto;
+
+public record GameCreatedDTO(Long gameId) {
+	
+}
