+++ src/main/java/com/trymad/litechess_monolith/websocket/api/dto/MoveRequest.java
@@ -1,4 +1,4 @@
-package com.trymad.litechess_monolith.websocket;
+package com.trymad.litechess_monolith.websocket.api.dto;
 
 import com.trymad.litechess_monolith.chessgame.api.model.ChessPiece;
