+++ src/main/java/com/trymad/litechess_monolith/websocket/api/event/MoveEvent.java
@@ -3,7 +3,7 @@ package com.trymad.litechess_monolith.websocket.api.event;
 import java.util.UUID;
 
 import com.trymad.litechess_monolith.shared.event.DomainEvent;
-import com.trymad.litechess_monolith.websocket.MoveRequest;
+import com.trymad.litechess_monolith.websocket.api.dto.MoveRequest;
 
 public record MoveEvent(
 	MoveRequest moveRequest, Long gameId, UUID playerId
