+++ src/main/java/com/trymad/litechess_monolith/websocket/api/event/QueueRegistryEvent.java
@@ -3,7 +3,7 @@ package com.trymad.litechess_monolith.websocket.api.event;
 import java.util.UUID;
 
 import com.trymad.litechess_monolith.shared.event.DomainEvent;
-import com.trymad.litechess_monolith.websocket.CreateGameRequest;
+import com.trymad.litechess_monolith.websocket.api.dto.CreateGameRequest;
 
 public record QueueRegistryEvent(CreateGameRequest request, UUID playerId) implements DomainEvent {
