+++ src/main/java/com/trymad/litechess_monolith/websocket/internal/controller/WebSocketController.java
@@ -1,10 +1,10 @@
-package com.trymad.litechess_monolith.websocket.internal;
+package com.trymad.litechess_monolith.websocket.internal.controller;
 
-import com.trymad.litechess_monolith.matchmaking.GameMakerService;
 import com.trymad.litechess_monolith.shared.event.EventPublisher;
-import com.trymad.litechess_monolith.websocket.CreateGameRequest;
-import com.trymad.litechess_monolith.websocket.MoveRequest;
+import com.trymad.litechess_monolith.websocket.api.dto.CreateGameRequest;
+import com.trymad.litechess_monolith.websocket.api.dto.MoveRequest;
 import com.trymad.litechess_monolith.websocket.api.event.MoveEvent;
+import com.trymad.litechess_monolith.websocket.api.event.QueueRegistryEvent;
 
 import lombok.RequiredArgsConstructor;
 
@@ -12,7 +12,6 @@ import java.security.Principal;
 import java.util.UUID;
 import java.util.logging.Logger;
 
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
@@ -25,20 +24,18 @@ public class WebSocketController {
     private final EventPublisher eventPublisher;
     private final Logger logger = Logger.getLogger("websocket");
 
-    @Qualifier("eventGameMakerService")
-    private final GameMakerService gameMakerService;
-
     @MessageMapping("game/{gameId}")
     // @SendTo("/topic/{gameId}/move")
     public void acceptMove(@Payload MoveRequest moveRequest, @DestinationVariable("gameId") Long gameId, Principal principal) {
-		final MoveEvent moveEvent = new MoveEvent(moveRequest, gameId, UUID.fromString(principal.getName()));
-		logger.info(moveRequest.toString());
-		eventPublisher.publish(moveEvent);
+      final MoveEvent moveEvent = new MoveEvent(moveRequest, gameId, UUID.fromString(principal.getName()));
+      logger.info(moveRequest.toString());
+      eventPublisher.publish(moveEvent);
     }
 
     @MessageMapping("matchmaking/queue")
     public void findGame(@Payload CreateGameRequest gameRequest, Principal principal) {
         System.out.println("request to sub");
-        gameMakerService.addInQueue(gameRequest, UUID.fromString(principal.getName()));
+        final QueueRegistryEvent event = new QueueRegistryEvent(gameRequest, UUID.fromString(principal.getName()));
+        eventPublisher.publish(event);
     }
 }
