+++ src/main/java/com/trymad/litechess_monolith/websocket/internal/listener/moveAccepted/SpringWebsocketMoveAcceptedEventListener.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 
 import com.trymad.litechess_monolith.chessgame.api.event.MoveAcceptedEvent;
 import com.trymad.litechess_monolith.shared.event.EventListener;
-import com.trymad.litechess_monolith.websocket.internal.GameMessageSender;
+import com.trymad.litechess_monolith.websocket.internal.service.GameMessageSender;
 
 import lombok.RequiredArgsConstructor;
