+++ src/main/java/com/trymad/litechess_monolith/websocket/internal/model/GameResultMessage.java
@@ -1,4 +1,4 @@
-package com.trymad.litechess_monolith.websocket.internal;
+package com.trymad.litechess_monolith.websocket.internal.model;
 
 import com.trymad.litechess_monolith.chessgame.api.model.ChessGameStatus;
