--- src/main/java/com/trymad/litechess_monolith/websocket/internal/service/gamemaker/QueueRegistryPublisher.java
@@ -1,9 +1,0 @@
-package com.trymad.litechess_monolith.websocket.internal.service.gamemaker;
-
-import com.trymad.litechess_monolith.websocket.api.event.QueueRegistryEvent;
-
-public interface QueueRegistryPublisher {
-	
-	public void publish(QueueRegistryEvent event);
-
-}
