--- src/main/java/com/trymad/litechess_monolith/websocket/internal/service/gamemaker/SpringQueueRegistryPublisher.java
@@ -1,21 +1,0 @@
-package com.trymad.litechess_monolith.websocket.internal.service.gamemaker;
-
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-import com.trymad.litechess_monolith.websocket.api.event.QueueRegistryEvent;
-
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class SpringQueueRegistryPublisher implements QueueRegistryPublisher {
-
-	private final ApplicationEventPublisher events;
-
-	@Override
-	public void publish(QueueRegistryEvent event) {
-		events.publishEvent(event);
-	}
-	
-}
